+++ tests/test_wiki_page.py
@@ -35,9 +35,6 @@ class WikiPageTests(PRAWTest):
         page.remove_editor(self.un)
         page.add_editor(self.un)
 
-        self.r.evict(self.r.config['wiki_page_settings'].format(
-                     subreddit=self.subreddit.display_name, page='index'))
-
         editors = page.get_settings()['editors']
         self.assertTrue(any(
             user.name.lower() == self.un.lower() for user in editors))